+++ scim2_cli/query.py
@@ -1,6 +1,5 @@
 import click
 from click import ClickException
-from scim2_client import SCIMClientError
 from scim2_models import ResourceType
 from scim2_models import Schema
 from scim2_models import SearchRequest
@@ -10,6 +9,7 @@ from sphinx_click.rst_to_ansi_formatter import make_rst_to_ansi_formatter
 from scim2_cli.utils import exception_to_click_error
 
 from .utils import DOC_URL
+from .utils import SCIM_EXCEPTIONS
 from .utils import formatted_payload
 
 
@@ -112,12 +112,12 @@ def query_cli(
         response = ctx.obj["client"].query(
             resource_type,
             id,
-            search_request=payload,
+            query_parameters=payload,
             check_request_payload=check_request_payload,
             raise_scim_errors=False,
         )
 
-    except SCIMClientError as scim_exc:
+    except SCIM_EXCEPTIONS as scim_exc:
         raise exception_to_click_error(scim_exc) from scim_exc
 
     payload = formatted_payload(response.model_dump(), indent)